jet-mysql/requirements.txt
requests
requests_toolbelt 
requests-oauthlib
python-dotenv
netifaces
